Fix math overflow

+++ src/internal/Rgbw64Color.h
@@ -232,7 +232,7 @@ private:
 
     inline static uint16_t _elementBrighten(uint16_t value, uint16_t ratio)
     {
-        uint16_t element = ((static_cast<uint32_t>(value) + 1) << 16) / (static_cast<uint32_t>(ratio) + 1);
+        uint32_t element = ((static_cast<uint32_t>(value) + 1) << 16) / (static_cast<uint32_t>(ratio) + 1);
 
         if (element > Max)
         {
